Revert "Feature/readme update"
