+++ .gitignore
@@ -1,6 +1,1 @@
 .env
-.python-version
-.venv
-__pycache__
-.pytest_cache
-project_proposal.md
