--- app/extensions/extensions.py
@@ -1,10 +1,0 @@
-from flask_login import LoginManager
-from peewee import SqliteDatabase
-
-
-# Flask Login
-login_manager = LoginManager()
-
-# Database Initalization
-db = SqliteDatabase(None)
-
