--- app/routes/templates.py
@@ -1,12 +1,0 @@
-from flask import Blueprint, render_template
-
-template_bp = Blueprint("templates", __name__)
-
-"""Render HTML Pages"""
-@template_bp.route("/")
-def index():
-    return render_template("index.html")
-
-@template_bp.route("/dashboard")
-def dashboard():
-    return render_template("dashboard.html")
