--- app/templates/index.html
@@ -1,48 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-
-<main>
-    <section class="introduction two-grid">
-        <div>
-            <p class="elms-sans-regular h1">For your Pantry Tracking Needs!</p>
-            <p class="elms-sans-regular h2">Track your pantry items before they disappear!</p>
-        </div>
-        <div class="container introduction-title">
-            <h1 class="elms-sans-regular h1">Snack Stash Storeroom</h1>
-        </div>
-    </section>
-
-    <div class="spacer-large"></div>
-
-    <h2 class="elms-sans-regular h1">Features</h2>
-
-    <section class="two-grid features">
-        <div>
-            <h2 class="elms-sans-regular h2">Item Tracking</h2>
-            <p class="quicksand-regular">
-                Snack Stash Storeroom is first and foremost a tracking app. <br>
-                Use this application to track expiration dates, item quantities, and your recipes!
-            </p>
-        </div>
-        <div class="spacer-medium features">
-            <img src="{{ url_for('static', filename='images/pantry-icon.png') }}" alt="img">
-        </div>
-        <div>
-            <h2 class="elms-sans-regular h2">Recipe Generation</h2>
-            <p class="quicksand-regular">
-                Users can generate recipes by providing an image in the "Recipes" section of their pantry list. <br>
-                These recipes are simple, and meant to give users an idea on how they want to create a new recipe. <br>
-                Suggestions can also be given to develop a recipe based on ingredients on a user's pantry.
-            </p><br><br><br><br><br><br><br>                    
-        </div>
-        <div class="spacer-medium features">
-            <img src="{{ url_for('static', filename='images/recipe-icon.png') }}" alt="img">
-        </div>
-    </section>
-</main>
-
-{% endblock %}
-
-{% block scripts %}
-<script src="{{ url_for('static', filename='script.js') }}"></script>
-{% endblock %}
