--- app/templates/layout.html
@@ -1,93 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-    <head>
-        <meta charset="UTF-8">
-        <meta name="viewport" content="width=device-width, initial-scale=1.0">
-        <title>Home</title>
-        <link rel="stylesheet" href="{{ url_for('static', filename='styles.css') }}">
-        <link rel="icon" type="image/x-icon" href="/favicon.ico">
-        <!-- Imported Links -->
-        <link rel="preconnect" href="https://fonts.googleapis.com">
-        <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
-        <link 
-        href="https://fonts.googleapis.com/css2?family=Elms+Sans:ital,wght@0,100..900;1,100..900&family=Quicksand:wght@300..700&display=swap" 
-        rel="stylesheet">
-    </head>
-    <body>
-        <nav>
-            <img src="" class="elms-sans-regular" alt="Snack Stash Storeroom">
-            <ul class="nav-links">
-                <li><a href="#" class="quicksand-regular">Home</a></li>
-                <li><a href="#" class="quicksand-regular">About</a></li>
-                <li><a href="#" class="quicksand-regular">Contacts</a></li>
-            </ul>
-
-            {% if current_user.is_authenticated %}
-
-            <ul class="login-links">
-                <li class="quicksand-regular"><p>{{ current_user.username }}</p></li>
-                <li>
-                    <form method="POST" action="{{ url_for('auth.logout') }}">
-                        <button type="submit" class="quicksand-regular">Logout</button>
-                    </form>
-                </li>
-            </ul>
-
-            {% else %}
-                <div id="overlay" class="overlay"></div>
-
-                <ul class="login-links">
-                    <li>
-                        <button type="button" id="login-btn" class="quicksand-regular">Login</button>
-                    </li>
-                </ul>
-
-                <div id="login-popup" class="popup-login">
-                
-                <button id="close-btn" class="close-btn">&times;</button>
-
-                <form id="login-form" method="post" action="{{ url_for('auth.login') }}">
-                    <h3 class="elms-sans-regular h1">Login</h3>
-                    <br>
-                    <p class="elms-sans-regular">Username:</p>
-                    <input type="text" name="username" 
-                    placeholder="Username" class="quicksand-regular" required>
-                    <p class="elms-sans-regular">Password:</p>
-                    <div class="password-wrapper">
-                        <input id="password-field" type="password" name="password" 
-                        placeholder="Password" class="quicksand-regular" required>
-                        <span id="toggle-password" class="toggle-eye elms-sans-regular">Show</span>
-                    </div>
-                    <div id="login-error" class="error-login quicksand-regular">
-                        <p id="error-text" class="quicksand-regular"></p>
-                    </div>
-                    <br>
-                    <button type="submit" class="quicksand-regular">Confirm</button>
-                </form>
-            </div>
-
-            {% endif %}
-        </nav>
-
-        {% block content %}
-        {% endblock %}
-
-        <footer>
-            <div class="footer-connection">
-                <img src="" alt="Discord" class="elms-sans-regular">
-            </div>
-            <div class="footer-message">
-
-            </div>
-            <ul class="footer-nav">
-                <li class="elms-sans-regular"><h1>Navigate</h1></li>
-                <li><a href="#" class="quicksand-regular">Home</a></li>
-                <li><a href="#" class="quicksand-regular">About</a></li>
-                <li><a href="#" class="quicksand-regular">Contacts</a></li>
-            </ul>
-        </footer>
-
-        {% block scripts %}
-        {% endblock %}
-    </body>
-</html>
