--- pyproject.toml
@@ -1,18 +1,0 @@
-[project]
-name = "acit-2911"
-version = "0.1.0"
-description = "Add your description here"
-readme = "README.md"
-requires-python = ">=3.14"
-dependencies = [
-    "flask>=3.1.3",
-    "flask-login>=0.6.3",
-    "peewee>=4.0.5",
-    "python-dotenv>=1.2.2",
-    "requests>=2.33.1",
-]
-
-[dependency-groups]
-dev = [
-    "pytest>=9.0.3",
-]
